# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ElytraCategory = typeof ElytraCategory[keyof typeof ElytraCategory];


export const ElytraCategory = {
  elytra: 'elytra',
} as const;

export type TimeRange = typeof TimeRange[keyof typeof TimeRange];


export const TimeRange = {
  five_minutes: 'five_minutes',
  hour: 'hour',
  today: 'today',
  seven_days: 'seven_days',
  thirty_days: 'thirty_days',
  ninety_days: 'ninety_days',
  one_year: 'one_year',
  all_time: 'all_time',
} as const;

export interface ElytraStats {
  category: ElytraCategory;
  /** @nullable */
  lowest: number | null;
  /** @nullable */
  highest: number | null;
  /** @nullable */
  average: number | null;
  /** @nullable */
  median: number | null;
  activeListings: number;
  /** @nullable */
  priceChange: number | null;
  currency: string;
}

export interface ApiStatus {
  connected: boolean;
  /** @nullable */
  lastUpdated: string | null;
  requestsInWindow: number;
  requestLimit: number;
  message: string;
}

export interface ElytraDashboard {
  stats: ElytraStats[];
  api: ApiStatus;
  qualifyingListings: number;
  generatedAt: string;
}

export interface PriceObservation {
  id: number;
  category: ElytraCategory;
  timestamp: string;
  price: number;
  open: number;
  high: number;
  low: number;
  close: number;
  /** @nullable */
  priceChange: number | null;
  sampleSize: number;
  observationCount: number;
}

export interface ElytraListing {
  id: string;
  /** @nullable */
  itemId: string | null;
  displayName: string;
  category: ElytraCategory;
  enchantments: string[];
  price: number;
  seller: string;
  /** @nullable */
  sellerUuid: string | null;
  quantity: number;
  /** @nullable */
  timeRemaining: string | null;
  collectedAt: string;
}

export interface ElytraTransaction {
  id: string;
  seller: string;
  price: number;
  category: ElytraCategory;
  enchantments: string[];
  quantity: number;
  timestamp: string;
}

export type MarketAlertType = typeof MarketAlertType[keyof typeof MarketAlertType];


export const MarketAlertType = {
  massive_buy: 'massive_buy',
  massive_sell: 'massive_sell',
} as const;

export interface MarketAlert {
  id: number;
  type: MarketAlertType;
  category: ElytraCategory;
  affectedQuantity: number;
  /** @nullable */
  previousPrice: number | null;
  /** @nullable */
  currentPrice: number | null;
  /** @nullable */
  percentageChange: number | null;
  /** @nullable */
  estimatedValue: number | null;
  detectedAt: string;
}

export interface AnalyzeElytraListingBody {
  /** @minLength 1 */
  listingId: string;
}

export interface AnalyzeElytraMarketBody {
  range?: TimeRange;
}

export type ElytraAnalysisResponseRecommendation = typeof ElytraAnalysisResponseRecommendation[keyof typeof ElytraAnalysisResponseRecommendation];


export const ElytraAnalysisResponseRecommendation = {
  BUY: 'BUY',
  SELL: 'SELL',
  HOLD: 'HOLD',
} as const;

export type ElytraAnalysisResponseMarketContext = {
  /** @nullable */
  lowest: number | null;
  /** @nullable */
  median: number | null;
  /** @nullable */
  average: number | null;
  /** @nullable */
  priceChange: number | null;
  activeListings: number;
  recentPrices: number[];
};

export type ElytraAnalysisResponseUsage = {
  used: number;
  remaining: number;
  limit: number;
};

export interface ElytraAnalysisResponse {
  listing: ElytraListing;
  recommendation: ElytraAnalysisResponseRecommendation;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  summary: string;
  reasons: string[];
  risks: string[];
  marketContext: ElytraAnalysisResponseMarketContext;
  usage: ElytraAnalysisResponseUsage;
}

export type ElytraMarketAnalysisResponseRecommendation = typeof ElytraMarketAnalysisResponseRecommendation[keyof typeof ElytraMarketAnalysisResponseRecommendation];


export const ElytraMarketAnalysisResponseRecommendation = {
  YES: 'YES',
  NO: 'NO',
} as const;

export type ElytraMarketAnalysisResponseMarketContext = {
  /** @nullable */
  lowest: number | null;
  /** @nullable */
  median: number | null;
  /** @nullable */
  average: number | null;
  /** @nullable */
  priceChange: number | null;
  activeListings: number;
  recentPrices: number[];
};

export type ElytraMarketAnalysisResponseSource = {
  auctionPages: number[];
  historyRange: TimeRange;
};

export type ElytraMarketAnalysisResponseUsage = {
  used: number;
  remaining: number;
  limit: number;
};

export interface ElytraMarketAnalysisResponse {
  recommendation: ElytraMarketAnalysisResponseRecommendation;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  summary: string;
  reasons: string[];
  risks: string[];
  marketContext: ElytraMarketAnalysisResponseMarketContext;
  source: ElytraMarketAnalysisResponseSource;
  usage: ElytraMarketAnalysisResponseUsage;
}

export type CategoryParameter = ElytraCategory;

export type TimeRangeParameter = TimeRange;

export type ListingSortParameter = typeof ListingSortParameter[keyof typeof ListingSortParameter];


export const ListingSortParameter = {
  lowest: 'lowest',
  highest: 'highest',
  recent: 'recent',
} as const;

export type GetElytraHistoryParams = {
category?: CategoryParameter;
range?: TimeRangeParameter;
};

export type GetElytraListingsParams = {
category?: CategoryParameter;
sort?: ListingSortParameter;
};

export type GetElytraTransactionsParams = {
category?: CategoryParameter;
};

export type GetElytraAlertsParams = {
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
/**
 * Minimum number of units affected for an alert to be returned
 * @minimum 1
 * @maximum 100
 */
threshold?: number;
};