# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Returns current qualifying Elytra market statistics, API usage, and update status.
 * @summary Get the Elytra market dashboard
 */
export const GetElytraDashboardResponse = zod.object({
  "stats": zod.array(zod.object({
  "category": zod.enum(['elytra']),
  "lowest": zod.number().nullable(),
  "highest": zod.number().nullable(),
  "average": zod.number().nullable(),
  "median": zod.number().nullable(),
  "activeListings": zod.number(),
  "priceChange": zod.number().nullable(),
  "currency": zod.string()
})),
  "api": zod.object({
  "connected": zod.boolean(),
  "lastUpdated": zod.coerce.date().nullable(),
  "requestsInWindow": zod.number(),
  "requestLimit": zod.number(),
  "message": zod.string()
}),
  "qualifyingListings": zod.number(),
  "generatedAt": zod.coerce.date()
})


/**
 * @summary Get real Elytra price history
 */
export const GetElytraHistoryQueryParams = zod.object({
  "category": zod.enum(['elytra']).optional(),
  "range": zod.enum(['five_minutes', 'hour', 'today', 'seven_days', 'thirty_days', 'ninety_days', 'one_year', 'all_time']).optional()
})

export const GetElytraHistoryResponseItem = zod.object({
  "id": zod.number(),
  "category": zod.enum(['elytra']),
  "timestamp": zod.coerce.date(),
  "price": zod.number(),
  "open": zod.number(),
  "high": zod.number(),
  "low": zod.number(),
  "close": zod.number(),
  "priceChange": zod.number().nullable(),
  "sampleSize": zod.number(),
  "observationCount": zod.number()
})
export const GetElytraHistoryResponse = zod.array(GetElytraHistoryResponseItem)


/**
 * @summary Get current qualifying Elytra listings
 */
export const getElytraListingsQuerySortDefault = `lowest`;

export const GetElytraListingsQueryParams = zod.object({
  "category": zod.enum(['elytra']).optional(),
  "sort": zod.enum(['lowest', 'highest', 'recent']).default(getElytraListingsQuerySortDefault)
})

export const GetElytraListingsResponseItem = zod.object({
  "id": zod.string(),
  "itemId": zod.string().nullable(),
  "displayName": zod.string(),
  "category": zod.enum(['elytra']),
  "enchantments": zod.array(zod.string()),
  "price": zod.number(),
  "seller": zod.string(),
  "sellerUuid": zod.string().nullable(),
  "quantity": zod.number(),
  "timeRemaining": zod.string().nullable(),
  "collectedAt": zod.coerce.date()
})
export const GetElytraListingsResponse = zod.array(GetElytraListingsResponseItem)


/**
 * @summary Get recent qualifying Elytra observations
 */
export const GetElytraTransactionsQueryParams = zod.object({
  "category": zod.enum(['elytra']).optional()
})

export const GetElytraTransactionsResponseItem = zod.object({
  "id": zod.string(),
  "seller": zod.string(),
  "price": zod.number(),
  "category": zod.enum(['elytra']),
  "enchantments": zod.array(zod.string()),
  "quantity": zod.number(),
  "timestamp": zod.coerce.date()
})
export const GetElytraTransactionsResponse = zod.array(GetElytraTransactionsResponseItem)


/**
 * @summary Get real Elytra market alerts
 */
export const getElytraAlertsQueryLimitDefault = 10;
export const getElytraAlertsQueryLimitMax = 50;

export const getElytraAlertsQueryThresholdDefault = 10;
export const getElytraAlertsQueryThresholdMax = 100;



export const GetElytraAlertsQueryParams = zod.object({
  "limit": zod.coerce.number().min(1).max(getElytraAlertsQueryLimitMax).default(getElytraAlertsQueryLimitDefault),
  "threshold": zod.coerce.number().min(1).max(getElytraAlertsQueryThresholdMax).default(getElytraAlertsQueryThresholdDefault).describe('Minimum number of units affected for an alert to be returned')
})

export const GetElytraAlertsResponseItem = zod.object({
  "id": zod.number(),
  "type": zod.enum(['massive_buy', 'massive_sell']),
  "category": zod.enum(['elytra']),
  "affectedQuantity": zod.number(),
  "previousPrice": zod.number().nullable(),
  "currentPrice": zod.number().nullable(),
  "percentageChange": zod.number().nullable(),
  "estimatedValue": zod.number().nullable(),
  "detectedAt": zod.coerce.date()
})
export const GetElytraAlertsResponse = zod.array(GetElytraAlertsResponseItem)


/**
 * Analyzes one current Elytra listing against the stored market context.
 * @summary Ask Gemini for a listing recommendation
 */



export const AnalyzeElytraListingBody = zod.object({
  "listingId": zod.string().min(1)
})

export const analyzeElytraListingResponseConfidenceMin = 0;
export const analyzeElytraListingResponseConfidenceMax = 100;



export const AnalyzeElytraListingResponse = zod.object({
  "listing": zod.object({
  "id": zod.string(),
  "itemId": zod.string().nullable(),
  "displayName": zod.string(),
  "category": zod.enum(['elytra']),
  "enchantments": zod.array(zod.string()),
  "price": zod.number(),
  "seller": zod.string(),
  "sellerUuid": zod.string().nullable(),
  "quantity": zod.number(),
  "timeRemaining": zod.string().nullable(),
  "collectedAt": zod.coerce.date()
}),
  "recommendation": zod.enum(['BUY', 'SELL', 'HOLD']),
  "confidence": zod.number().min(analyzeElytraListingResponseConfidenceMin).max(analyzeElytraListingResponseConfidenceMax),
  "summary": zod.string(),
  "reasons": zod.array(zod.string()),
  "risks": zod.array(zod.string()),
  "marketContext": zod.object({
  "lowest": zod.number().nullable(),
  "median": zod.number().nullable(),
  "average": zod.number().nullable(),
  "priceChange": zod.number().nullable(),
  "activeListings": zod.number(),
  "recentPrices": zod.array(zod.number())
}),
  "usage": zod.object({
  "used": zod.number(),
  "remaining": zod.number(),
  "limit": zod.number()
})
})


/**
 * Sends DonutSMP auction pages one and two plus the selected stored price history range to Gemini.
 * @summary Ask Gemini for a market-level buy decision
 */
export const analyzeElytraMarketBodyRangeDefault = `hour`;

export const AnalyzeElytraMarketBody = zod.object({
  "range": zod.enum(['five_minutes', 'hour', 'today', 'seven_days', 'thirty_days', 'ninety_days', 'one_year', 'all_time']).default(analyzeElytraMarketBodyRangeDefault)
})

export const analyzeElytraMarketResponseConfidenceMin = 0;
export const analyzeElytraMarketResponseConfidenceMax = 100;



export const AnalyzeElytraMarketResponse = zod.object({
  "recommendation": zod.enum(['YES', 'NO']),
  "confidence": zod.number().min(analyzeElytraMarketResponseConfidenceMin).max(analyzeElytraMarketResponseConfidenceMax),
  "summary": zod.string(),
  "reasons": zod.array(zod.string()),
  "risks": zod.array(zod.string()),
  "marketContext": zod.object({
  "lowest": zod.number().nullable(),
  "median": zod.number().nullable(),
  "average": zod.number().nullable(),
  "priceChange": zod.number().nullable(),
  "activeListings": zod.number(),
  "recentPrices": zod.array(zod.number())
}),
  "source": zod.object({
  "auctionPages": zod.array(zod.number()),
  "historyRange": zod.enum(['five_minutes', 'hour', 'today', 'seven_days', 'thirty_days', 'ninety_days', 'one_year', 'all_time'])
}),
  "usage": zod.object({
  "used": zod.number(),
  "remaining": zod.number(),
  "limit": zod.number()
})
})