# lib/api-zod/src/generated/types/analyzeElytraMarketBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TimeRange } from './timeRange';

export interface AnalyzeElytraMarketBody {
  range?: TimeRange;
}